Breaking changes! Probably doesn't build.
This is a commit of code drafting, notes, and other miscellany
that I wrote many months before this commit.

I probably didn't want to commit this because it's probably shit,
but right now I just want a clean git-index so that I can rename
things and track my more recent work.